backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
pillow
django-cors-headers
